카테고리 디자인, 분자 컴포넌트 분리 (#187)
* fix: 분류 카테고리 버튼식 디자인으로 변경

* feat: 유저, 전체 카테고리 리스트

* fix: 중복 제거
