+++ src/components/molecules/card/category/CategorySortCard.tsx
@@ -17,21 +17,18 @@ const CategorySortCard = ({ categoryName }: CategorySortCardProps) => {
           pathname: router.pathname,
           query: { ...router.query, category: categoryName },
         }}
-        className={`underline decoration-${categoryName} text-sm font-semibold text-${categoryName} font-eng-sub-font-1`}
+        className={`text-${categoryName} p-2 text-sm font-medium rounded-md ring-1 bg-neutral-50  ring-neutral-300 font-eng-sub-font-2 hover:ring-neutral-400 dark:bg-neutral-900 dark:ring-neutral-50 hover:dark:ring-neutral-600 dark:text-neutral-50 hover:dark:text-neutral-500`}
       >
-        #
-        <span className="text-neutral-600 hover:text-neutral-800 dark:text-neutral-50">
-          {categoryName}
-        </span>
+        # <span className="text-neutral-800">{categoryName}</span>
       </LinkTo>
     );
   }
   return (
     <LinkTo
       href={`${router.asPath.split("?")[0]}`}
-      className={`underline decoration-${categoryName} text-sm font-semibold text-${categoryName} font-eng-sub-font-1`}
+      className={`p-2 text-sm font-medium rounded-md ring-1 bg-neutral-50 text-${categoryName} ring-neutral-300 hover:ring-neutral-400 font-eng-sub-font-2 dark:bg-neutral-900`}
     >
-      #<span className="">{categoryName}</span>
+      # {categoryName}
     </LinkTo>
   );
 };
