+++ src/components/organisms/card-list/category/index.ts
@@ -1,0 +1,4 @@
+import AllCategoryCardList from "@Components/organisms/card-list/category/AllCategoryCardList";
+import UserCategoryCardList from "@Components/organisms/card-list/category/UserCategoryCardList";
+
+export { AllCategoryCardList, UserCategoryCardList };
